add support mpeg4 rtsp streams (355)

+++ upload/upload.c
@@ -8,11 +8,15 @@
 #include <signal.h>
 #include <pthread.h>
 #include <errno.h>
+#include <sys/wait.h>
 
 #define BASE_DIR		"/media/sda1/MVR"
 #define FTP_SERVER		"192.168.1.2"
 #define FTP_USER		"tvhelp"
 #define FTP_PASSWD		"tvhelp"
+#define TAG				"UPLOAD:"
+//#define WITH_CONTROL_THREAD
+//#define WITH_FTP_FORK
 
 int run;
 pthread_t pth_control;
@@ -21,36 +25,68 @@ char ftpServer[256];
 char ftpUser[256];
 char ftpPasswd[256];
 char oldDir[256];
-
+#ifdef WITH_FTP_FORK
+pid_t ftpPid = 0;
+#endif
 void sig_handler(int signum);
+#ifdef WITH_CONTROL_THREAD
 void* control_func (void *arg);
+#endif
 
 void print_help(char *argv)
 {
 	printf("Usage %s <base_dir> <ftp_server> <ftp_user> <ftp_password>\n",argv);
 }
 
-void uploadToFTPServer(char *file)
+int uploadToFTPServer(char *file)
 {
 	int ret;
 	char cmd[1024];
 //	printf("%s %s\n",__func__, file);
-	
-	sprintf(cmd, "ftpput -u %s -p %s %s %s", ftpUser, ftpPasswd, ftpServer, file);
-	printf("%s\n",cmd);
-	ret = system(cmd);
-	//TODO check ftpput return remove file
-	if(ret != -1 && ret != 127)
+#ifdef WITH_FTP_FORK
+	ret = kill(ftpPid, 0);
+//	printf("kill pid %d with 0 return %d\n", ftpPid, ret);
+	if(ret == -1 || ftpPid == 0)//previous ftpput complete
 	{
-		if(ret == EXIT_SUCCESS)
+		ftpPid = fork();
+		if(ftpPid > 0)
 		{
-			sprintf(cmd, "rm -f %s", file);
-			printf("%s\n",cmd);
-			system(cmd);
+			printf("%s create child process with PID %d\n", TAG, ftpPid);
 		}
-		else if(ret == EXIT_FAILURE)
-			printf("ERROR!!! Can't upload %s to ftp %s\n", file, ftpServer);
+		if(ftpPid == -1)
+		{
+			perror("fork"); /* произошла ошибка */
+			exit(1); /*выход из родительского процесса*/
+		}
+		if(ftpPid == 0)
+		{
+			printf("Start of child ftp process\n");
+#endif
+			sprintf(cmd, "ftpput -u %s -p %s %s %s", ftpUser, ftpPasswd, ftpServer, file);
+			printf("%s ftpput %s\n",TAG, file);
+			ret = system(cmd);
+			//TODO check ftpput return remove file
+			if(ret != 0)
+				printf("\t\nftpput return ERROR %d\n", ret);
+			if(ret == 256) exit(1);
+			if(ret != -1 && ret != 127)
+			{
+				if(ret == EXIT_SUCCESS)
+				{
+					sprintf(cmd, "rm -f %s", file);
+					printf("%s remove %s\n",TAG, file);
+					system(cmd);
+				}
+				else if(ret == EXIT_FAILURE)
+					printf("ERROR!!! Can't upload %s to ftp %s\n", file, ftpServer);
+			}
+#ifdef WITH_FTP_FORK
+			printf("End of child ftp process\n");
+			return 0;
+		}
 	}
+#endif
+	return 0;
 }
 
 void removeDir(char *path)
@@ -81,7 +117,7 @@ int searchOldestFile(char *path)
 	int year, month, day;
 	int currDay, currDir;
 	currDay = currDir = 0;
-//	printf("%s\n",__func__);
+//	printf("%s in %s\n",__func__, path);
 	
 	dp = opendir(path);
 	if( dp != NULL)
@@ -176,12 +212,13 @@ int searchOldestDir(char *path, char *dir)
 	return -ENOENT;
 }
 
-//Search in BASE_DIR old files and upload them on ftp server
+//Search in baseDir old files and upload them on ftp server
 int main (int argc, char *argv[])
 {
 	int ret, oldestDirEmpty;
 	char dir[256];
 	char camera[256];
+	
 	DIR *dp;
 	struct dirent *ep;   
 	
@@ -205,21 +242,28 @@ int main (int argc, char *argv[])
 
 	printf("baseDir (%s) FTP: server (%s), user (%s), passwd (%s)\n", baseDir, ftpServer, ftpUser, ftpPasswd);
 	
-	signal(SIGINT, sig_handler);
+//	sprintf(cmd, "ping %s", ftpServer);
+//	system(cmd);
+	if (signal(SIGINT, sig_handler) == SIG_ERR)
+		printf("\t\n%s can't catch SIGINT\n", TAG);
+#ifdef WITH_FTP_FORK
+	signal(SIGCHLD, sig_handler);
+#endif	
 	run = 1;
-	
+#ifdef WITH_CONTROL_THREAD	
 	ret = pthread_create(&pth_control, NULL, &control_func, argv[0]);
 	if (ret != 0)
 	{
-		printf("\ncan't create control thread :[%s]", strerror(ret));
+		printf("\n%s can't create control thread :[%s]", TAG, strerror(ret));
 		return -1;
 	}
 	else
-	    printf("\n Control thread created successfully\n");
-
+	    printf("\n%s control thread created successfully\n", TAG);
+#endif
 	while(run)
 	{
 		//Search for oldest day
+//		printf("%s while run\n", TAG);
 		dir[0] = '\0';
 		ret = searchOldestDir(baseDir, dir);
 		if(ret == 0)
@@ -230,7 +274,7 @@ int main (int argc, char *argv[])
 				camera[0] = 0;
 				oldestDirEmpty = 0;//for check if any camera directory exist in oldest dir
 				//In each camera folder search oldest file
-				while ((ep = readdir (dp)))
+				while ((ep = readdir (dp)) && run)
 				{
 					int retVal;
 					if( (0 == strcmp(ep->d_name,".") )|| ( 0 == strcmp(ep->d_name,"..")) ) continue;
@@ -263,20 +307,44 @@ int main (int argc, char *argv[])
 			run = 0;
 //			break;
 		}
-		usleep(100000);
+		usleep(10000);
 	}
+#ifdef WITH_CONTROL_THREAD
 	if(pth_control)
 	    pthread_join(pth_control, NULL);
+#endif
+//	sprintf(cmd, "killall -9 ftpput");
+//	system(cmd);
 
+	printf("\t\n%s STOPPED.\n", TAG);
 	return 0;
 }
 
 void sig_handler(int signum)
 {
-    printf("Received signal %d\n", signum);
-    run = 0;
+#ifdef WITH_FTP_FORK
+	char cmd[256];
+	int wstatus;
+	if(signum == SIGCHLD)	
+	{
+//	    printf("\t\n%s Received signal %d\n", TAG, signum);
+		wait(&wstatus);
+//	    printf("\t\n%s wait return %d\n", TAG, wstatus);
+	}
+#endif
+	if(signum == SIGINT)
+	{
+	    printf("\t\n%s Received signal %d\n", TAG, signum);
+#ifdef WITH_FTP_FORK
+		sprintf(cmd, "kill %d", ftpPid);
+		printf("%s %s\n", TAG, cmd);
+		system(cmd);
+#endif
+    	run = 0;
+    }
 }
 
+#ifdef WITH_CONTROL_THREAD
 void* control_func (void *arg)
 {
 	char c;
@@ -291,8 +359,9 @@ void* control_func (void *arg)
 				run = 0;
 				break;
 		}
-		usleep(100000);
+		usleep(10000);
 	}
+	printf("\t\n%s control thread STOPPED.\n", TAG);
 	return arg;
 }
-
+#endif
